+++ apps/users/admin.py
@@ -5,3 +5,9 @@ Siege. All rights reserved
 :copyright: (c) 2022-present Siege Team
 :author: Siege Team
 """
+
+from django.contrib import admin
+
+from apps.users.models import User
+
+admin.site.register(User)