+++ server/wsgi.py
@@ -1,16 +1,15 @@
 """
-WSGI config for server project.
+Siege. All rights reserved
+~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-It exposes the WSGI callable as a module-level variable named ``application``.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/4.1/howto/deployment/wsgi/
+:copyright: (c) 2022-present Siege Team
+:author: Siege Team
 """
 
 import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'server.settings')
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "server.settings")
 
 application = get_wsgi_application()